PLATTER-296 - Add weight scale quantity (#153)

+++ backend/app/graphql/types/ticket_item_attributes.rb
@@ -2,7 +2,7 @@
 
 class Types::TicketItemAttributes < Types::BaseInputObject
   argument :note, String, required: false
-  argument :quantity, Integer, required: false
+  argument :quantity, Float, required: false
   argument :status, enum("TicketItemStatusAttribute", TicketItem.statuses.keys), required: false
 
   argument :id, ID, required: false # Required for update
